Typos

+++ src/mc/explo/odpor/Execution.cpp
@@ -15,6 +15,7 @@
 #include "src/mc/transition/TransitionActor.hpp"
 
 #include <algorithm>
+#include <cstddef>
 #include <exception>
 #include <iterator>
 #include <limits>
@@ -482,7 +483,7 @@ std::optional<PartialExecution> Execution::get_odpor_extension_from(EventHandle 
 
     // If we are cutting exploration at some arbitrary depth, we might not find what we are looking for
     // It's okay, we just don't cut any exploration opportunity in that case
-    if (next_transition_aid == this->end() and _sg_mc_max_depth <= this->size())
+    if (next_transition_aid == this->end() and (size_t) _sg_mc_max_depth <= this->size())
       continue;
 
     xbt_assert(next_transition_aid != this->end(),
